test: isolate email identity resolution
